requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.7.4
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.8
dnspython==2.6.1
eventlet==0.36.1
exceptiongroup==1.2.2
fastapi==0.112.0
filelock==3.15.4
Flask==3.0.3
Flask-SocketIO==5.3.6
fonttools==4.53.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
idna==3.7
imageio==2.34.2
importlib_metadata==8.0.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1
numpy==2.0.1
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
pyav==12.0.5
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.3
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.5
virtualenv==20.26.3
websockets==12.0
Werkzeug==3.0.3
wsproto==1.2.0
zipp==3.19.2
